requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
psutil>=5.8.0

# Monitoring and logging (optional but recommended)
tensorboard>=2.10.0

# For CUDA support (if available)
# Note: CUDA toolkit should be installed separately via conda or system package manager

python-dotenv
requests